Add initial probe saving
