+++ elk/utils/wandb_utils.py
@@ -1,36 +1,7 @@
-from argparse import Namespace
-from copy import deepcopy
+from pathlib import Path
 from typing import Optional, Tuple
 
 import wandb
-from elk.evaluation.evaluate import Eval
-from elk.training.sweep import Sweep
-from elk.training.train import Elicit
-
-
-def wandb_init_helper(
-    args: Namespace, project_name: str = "elk_test_experiment"
-) -> None:
-    """
-    Serializes args so they can be logged in wandb
-    and starts a run according to the command type.
-    """
-    if args.wandb_tracking:
-        if isinstance(args, Eval):
-            args_serialized = deepcopy(args)
-            args_serialized.out_dir = str(
-                args_serialized.out_dir
-            )  # .as_posix method would break on windows
-            args_serialized.source = str(args_serialized.source)
-            wandb.init(project=project_name, config=args_serialized, job_type="eval")
-        elif isinstance(args, Elicit):
-            wandb.init(project=project_name, config=args, job_type="train")
-        elif isinstance(args, Sweep):
-            wandb.init(
-                project=project_name, config=args, job_type="sweep", group="Sweep1"
-            )
-    else:
-        wandb.init(mode="disabled")
 
 
 def find_run_details(entity_name: str, run_name: str) -> Optional[Tuple[str, str]]:
@@ -63,3 +34,11 @@ def find_run_details(entity_name: str, run_name: str) -> Optional[Tuple[str, str
 
     # If no run found with the given name, return None
     return None
+
+
+def wandb_save_probes(out_dir: Path) -> None:
+    if wandb.run is not None:
+        for dir in ["lr_models", "reporters"]:
+            artifact = wandb.Artifact(dir, type="model")
+            artifact.add_dir(out_dir / dir)
+            wandb.run.log_artifact(artifact)
